feat: 新增武器系统与武器选择界面
- 新增 Cotton Candy Gun 和 Deboink 两款武器及其数据配置
- 新增 WeaponSelectionManager 武器选择管理器
- 新增 PlayerWeapons 玩家武器组件管理多武器装备
- 新增 WeaponPosition 武器挂载点组件
- 新增 ItemLevelColorHelper 道具等级颜色辅助类
- 重构 UI 预制体结构，优化武器选择界面
- 更新 WaveTransitionManager 过渡动画管理
- 调整 PropType 道具类型定义

影响范围：武器系统、UI系统、玩家控制系统
修改文件：20+ 个脚本和预制体文件

+++ Assets/Scripts/Player/Bullet.cs
@@ -8,7 +8,7 @@ public class Bullet : MonoBehaviour
     private Rigidbody2D _rb;
     private Collider2D _collider;
     protected float _moveSpeed = 5;
-    protected int _damage;
+    protected float _damage;
     protected bool _isCritical;
     [SerializeField]protected LayerMask targetsLayerMask;
 
@@ -18,7 +18,7 @@ public class Bullet : MonoBehaviour
         _rb = GetComponent<Rigidbody2D>();
     }
 
-    public void Shoot(Vector2 direction, int damage,bool isCritical)
+    public void Shoot(Vector2 direction, float damage,bool isCritical)
     {
         _damage = damage;
         _isCritical = isCritical;
